+++ playground/pages/index.vue
@@ -1,7 +1,13 @@
 <template>
-  <div>
+  <div class="m-5">
     <h1>{{ title }}</h1>
     <h3>You are currently logged {{ loggedIn ? 'in' : 'out' }}</h3>
+    <nuxt-link
+      v-if="loggedIn"
+      to="/profile"
+    >
+      My super-secret profile page
+    </nuxt-link>
   </div>
 </template>
 
@@ -12,7 +18,7 @@ const { user, loggedIn } = useJwtAuth()
 
 const title = computed(() => {
   if (loggedIn) {
-    return 'Hello, ' + user?.name
+    return 'Welcome back, ' + user?.name
   } else {
     return 'Hello, handsome stranger'
   }